added refresh command, faker image provider, slug trait

+++ database/migrations/2022_10_07_182946_create_brands_table.php
@@ -11,12 +11,13 @@ return new class extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('brands', static function (Blueprint $table) {
             $table->id();
-            $table->text('title');
-            $table->text('description');
+            $table->string('slug')->unique();
+            $table->string('title');
+            $table->string('thumbnail')->nullable();
             $table->timestamps();
         });
     }
@@ -26,8 +27,10 @@ return new class extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
-        Schema::dropIfExists('brands');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('brands');
+        }
     }
 };
